requirements.txt
customtkinter
networkx
mplcursors
matplotlib
plotly
kaleido
Pillow
gitpython